+++ pandas_market_calendars/calendar_registry.py
@@ -3,18 +3,23 @@ from .calendars.asx import ASXExchangeCalendar
 from .calendars.bmf import BMFExchangeCalendar
 from .calendars.bse import BSEExchangeCalendar
 from .calendars.cboe import CFEExchangeCalendar
-from .calendars.cme import \
-    CMEEquityExchangeCalendar, \
-    CMEBondExchangeCalendar
+from .calendars.cme import CMEEquityExchangeCalendar, CMEBondExchangeCalendar
 from .calendars.cme_globex_base import CMEGlobexBaseExchangeCalendar
-from .calendars.cme_globex_agriculture import CMEGlobexAgricultureExchangeCalendar
+from .calendars.cme_globex_agriculture import (
+    CMEGlobexAgricultureExchangeCalendar,
+    CMEGlobexLivestockExchangeCalendar,
+    CMEGlobexGrainsAndOilseedsExchangeCalendar,
+)
 from .calendars.cme_globex_crypto import CMEGlobexCryptoExchangeCalendar
-from .calendars.cme_globex_energy_and_metals import CMEGlobexEnergyAndMetalsExchangeCalendar
+from .calendars.cme_globex_energy_and_metals import (
+    CMEGlobexEnergyAndMetalsExchangeCalendar,
+)
 from .calendars.cme_globex_equities import CMEGlobexEquitiesExchangeCalendar
 from .calendars.cme_globex_fx import CMEGlobexFXExchangeCalendar
 from .calendars.cme_globex_fixed_income import CMEGlobexFixedIncomeCalendar
 from .calendars.dukascopy_fx import DukascopyForexExchangeCalendar
 from .calendars.eurex import EUREXExchangeCalendar
+from .calendars.eurex_fixed_income import EUREXFixedIncomeCalendar
 from .calendars.hkex import HKEXExchangeCalendar
 from .calendars.ib_fx import IBForexExchangeCalendar
 from .calendars.ice import ICEExchangeCalendar
@@ -23,7 +28,11 @@ from .calendars.jpx import JPXExchangeCalendar
 from .calendars.lse import LSEExchangeCalendar
 from .calendars.nyse import NYSEExchangeCalendar
 from .calendars.ose import OSEExchangeCalendar
-from .calendars.sifma import SIFMAUSExchangeCalendar, SIFMAUKExchangeCalendar, SIFMAJPExchangeCalendar
+from .calendars.sifma import (
+    SIFMAUSExchangeCalendar,
+    SIFMAUKExchangeCalendar,
+    SIFMAJPExchangeCalendar,
+)
 from .calendars.six import SIXExchangeCalendar
 from .calendars.sse import SSEExchangeCalendar
 from .calendars.tase import TASEExchangeCalendar