+++ pandas_market_calendars/calendars/cme_globex_agriculture.py
@@ -14,15 +14,27 @@
 # limitations under the License.
 
 from abc import abstractmethod
-from .cme_globex_base import CMEGlobexBaseExchangeCalendar
-
 from datetime import time
 
-from pandas.tseries.holiday import AbstractHolidayCalendar, GoodFriday, USLaborDay, USPresidentsDay, USThanksgivingDay
+from pandas.tseries.holiday import (
+    AbstractHolidayCalendar,
+    GoodFriday,
+    USLaborDay,
+    USPresidentsDay,
+    USThanksgivingDay,
+)
 
-from pandas_market_calendars.holidays.us import (Christmas, ChristmasEveBefore1993, ChristmasEveInOrAfter1993, USBlackFridayInOrAfter1993,
-                                                 USIndependenceDay, USMartinLutherKingJrAfter1998, USMemorialDay,
-                                                 USNewYearsDay)
+from pandas_market_calendars.holidays.us import (
+    Christmas,
+    ChristmasEveBefore1993,
+    ChristmasEveInOrAfter1993,
+    USBlackFridayInOrAfter1993,
+    USIndependenceDay,
+    USMartinLutherKingJrAfter1998,
+    USMemorialDay,
+    USNewYearsDay,
+)
+from .cme_globex_base import CMEGlobexBaseExchangeCalendar
 
 
 class CMEGlobexAgricultureExchangeCalendar(CMEGlobexBaseExchangeCalendar):
@@ -37,10 +49,10 @@ class CMEGlobexAgricultureExchangeCalendar(CMEGlobexBaseExchangeCalendar):
     - Lumber and Softs
 
 
-     """
+    """
 
     @property
-    @abstractmethod 
+    @abstractmethod
     def name(self):
         """
         Name of the market
@@ -60,11 +72,18 @@ class CMEGlobexLivestockExchangeCalendar(CMEGlobexAgricultureExchangeCalendar):
     https://www.cmegroup.com/markets/agriculture/livestock/live-cattle.contractSpecs.html
     Monday - Friday: 8:30 a.m. - 1:05 p.m. CT
     """
-    aliases = ['CMEGlobex_Livestock', 'CMEGlobex_Live_Cattle', 'CMEGlobex_Feeder_Cattle', 'CMEGlobex_Lean_Hog', 'CMEGlobex_Port_Cutout']
+
+    aliases = [
+        "CMEGlobex_Livestock",
+        "CMEGlobex_Live_Cattle",
+        "CMEGlobex_Feeder_Cattle",
+        "CMEGlobex_Lean_Hog",
+        "CMEGlobex_Port_Cutout",
+    ]
 
     regular_market_times = {
         "market_open": ((None, time(8, 30)),),
-        "market_close": ((None, time(13, 5)),)
+        "market_close": ((None, time(13, 5)),),
     }
 
     @property
@@ -73,17 +92,19 @@ class CMEGlobexLivestockExchangeCalendar(CMEGlobexAgricultureExchangeCalendar):
 
     @property
     def regular_holidays(self):
-        return AbstractHolidayCalendar(rules=[
-            USNewYearsDay,
-            USMartinLutherKingJrAfter1998,
-            USPresidentsDay,
-            GoodFriday,
-            USMemorialDay,
-            USIndependenceDay,
-            USLaborDay,
-            USThanksgivingDay,
-            Christmas,
-        ])
+        return AbstractHolidayCalendar(
+            rules=[
+                USNewYearsDay,
+                USMartinLutherKingJrAfter1998,
+                USPresidentsDay,
+                GoodFriday,
+                USMemorialDay,
+                USIndependenceDay,
+                USLaborDay,
+                USThanksgivingDay,
+                Christmas,
+            ]
+        )
 
     # @property
     # def adhoc_holidays(self):
@@ -91,13 +112,61 @@ class CMEGlobexLivestockExchangeCalendar(CMEGlobexAgricultureExchangeCalendar):
 
     @property
     def special_closes(self):
-        return [(
-            time(12, 5),
-            AbstractHolidayCalendar(rules=[
-                USBlackFridayInOrAfter1993,
-                ChristmasEveBefore1993,
-                ChristmasEveInOrAfter1993,
-            ])
-        )]
+        return [
+            (
+                time(12, 5),
+                AbstractHolidayCalendar(
+                    rules=[
+                        USBlackFridayInOrAfter1993,
+                        ChristmasEveBefore1993,
+                        ChristmasEveInOrAfter1993,
+                    ]
+                ),
+            )
+        ]
 
 
+class CMEGlobexGrainsAndOilseedsExchangeCalendar(CMEGlobexAgricultureExchangeCalendar):
+    """
+    Exchange calendar for CME for Grains & Oilseeds
+
+    https://www.cmegroup.com/trading/agricultural/grain-and-oilseed.html
+
+    GLOBEX Trading Times
+    https://www.cmegroup.com/markets/agriculture/oilseeds/soybean.contractSpecs.html
+    https://www.cmegroup.com/markets/agriculture/grains/corn.contractSpecs.html
+    https://www.cmegroup.com/markets/agriculture/grains/wheat.contractSpecs.html
+    Sunday - Friday: 7:00 p.m. - 7:45 a.m. CT and Monday - Friday: 8:30 a.m. - 1:20 p.m. CT
+    """
+
+    aliases = [
+        "CMEGlobex_Grains",
+        "CMEGlobex_Oilseeds",
+    ]
+
+    regular_market_times = {
+        "market_open": ((None, time(19), -1),),  # offset by -1 day
+        "market_close": ((None, time(13, 20)),),
+        "break_start": ((None, time(7, 45)),),
+        "break_end": ((None, time(8, 30)),),
+    }
+
+    @property
+    def name(self):
+        return "CMEGlobex_GrainsAndOilseeds"
+
+    @property
+    def regular_holidays(self):
+        return AbstractHolidayCalendar(
+            rules=[
+                USNewYearsDay,
+                USMartinLutherKingJrAfter1998,
+                USPresidentsDay,
+                GoodFriday,
+                USMemorialDay,
+                USIndependenceDay,
+                USLaborDay,
+                USThanksgivingDay,
+                Christmas,
+            ]
+        )