+++ pandas_market_calendars/calendars/eurex_fixed_income.py
@@ -2,12 +2,9 @@ from datetime import time
 
 from pandas.tseries.holiday import (
     AbstractHolidayCalendar,
-    Day,
-    Easter,
     EasterMonday,
     GoodFriday,
     Holiday,
-    previous_friday,
 )
 from pytz import timezone
 
@@ -20,25 +17,11 @@ from pandas_market_calendars.market_calendar import (
     WEDNESDAY,
 )
 
-# New Year's Eve
-NewYearsEve = Holiday(
-    "New Year's Eve",
-    month=12,
-    day=31,
-    observance=previous_friday,
-)
 # New Year's Day
-NewYearsDay = Holiday(
+EUREXNewYearsDay = Holiday(
     "New Year's Day",
     month=1,
     day=1,
-    days_of_week=(MONDAY, TUESDAY, WEDNESDAY, THURSDAY, FRIDAY),
-)
-# Berthold's Day
-BertholdsDay = Holiday(
-    "Berthold's Day",
-    month=1,
-    day=2,
     days_of_week=(MONDAY, TUESDAY, WEDNESDAY, THURSDAY, FRIDAY),
 )
 # Early May bank holiday
@@ -48,34 +31,12 @@ MayBank = Holiday(
     day=1,
     days_of_week=(MONDAY, TUESDAY, WEDNESDAY, THURSDAY, FRIDAY),
 )
-# Ascension Day (Auffahrt)
-AscensionDay = Holiday(
-    "Ascension Day",
-    month=1,
-    day=1,
-    offset=[Easter(), Day(39)],
-    days_of_week=(THURSDAY,),
-)
-# Pentecost Day (Pfingstmontag)
-PentecostMonday = Holiday(
-    "Pentecost Monday",
-    month=1,
-    day=1,
-    offset=[Easter(), Day(50)],
-    days_of_week=(MONDAY,),
-)
-# Swiss National Day
-SwissNationalDay = Holiday(
-    "Swiss National Day",
-    month=8,
-    day=1,
-    days_of_week=(MONDAY, TUESDAY, WEDNESDAY, THURSDAY, FRIDAY),
-)
 # Christmas Eve
 ChristmasEve = Holiday(
     "Christmas Eve",
     month=12,
     day=24,
+    days_of_week=(MONDAY, TUESDAY, WEDNESDAY, THURSDAY, FRIDAY),
 )
 # Christmas
 Christmas = Holiday(
@@ -89,41 +50,46 @@ BoxingDay = Holiday(
     "Boxing Day",
     month=12,
     day=26,
+    days_of_week=(MONDAY, TUESDAY, WEDNESDAY, THURSDAY, FRIDAY),
+)
+# New Year's Eve
+NewYearsEve = Holiday(
+    "New Year's Eve",
+    month=12,
+    day=31,
+    days_of_week=(MONDAY, TUESDAY, WEDNESDAY, THURSDAY, FRIDAY),
 )
 
 
-class SIXExchangeCalendar(MarketCalendar):
+class EUREXFixedIncomeCalendar(MarketCalendar):
     """
-    Exchange calendar for SIX
-
+    Trading calendar available here:
+    https://www.eurex.com/resource/blob/3378814/910cf372738890f691bc1bfbccfd3aef/data/tradingcalendar_2023_en.pdf
     """
 
-    aliases = ["SIX"]
+    aliases = ["EUREX_Bond"]
+
     regular_market_times = {
-        "market_open": ((None, time(9)),),
-        "market_close": ((None, time(17, 30)),),
+        "market_open": ((None, time(1, 10)), ("2018-12-10", time(8, 0))),
+        "market_close": ((None, time(22)),),
     }
 
     @property
     def name(self):
-        return "SIX"
+        return "EUREX_Bond"
 
     @property
     def tz(self):
-        return timezone("Europe/Zurich")
+        return timezone("Europe/Berlin")
 
     @property
     def regular_holidays(self):
         return AbstractHolidayCalendar(
             rules=[
-                NewYearsDay,
-                BertholdsDay,
+                EUREXNewYearsDay,
                 GoodFriday,
                 EasterMonday,
                 MayBank,
-                AscensionDay,
-                PentecostMonday,
-                SwissNationalDay,
                 ChristmasEve,
                 Christmas,
                 BoxingDay,