+++ pandas_market_calendars/calendars/mirror.py
@@ -4,8 +4,11 @@ Imported calendars from the exchange_calendars project
 GitHub: https://github.com/gerrymanoim/exchange_calendars
 """
 
-from pandas_market_calendars.market_calendar import MarketCalendar
 import exchange_calendars
+
+from pandas_market_calendars.market_calendar import MarketCalendar
+
+DAYMASKS = ["Mon", "Tue", "Wed", "Thu", "Fri", "Sat", "Sun"]
 
 
 class TradingCalendar(MarketCalendar):
@@ -17,6 +20,7 @@ class TradingCalendar(MarketCalendar):
     The initialization of calendars from exchange_calendars, is bypassed until the `.ec` property is used,
     which returns the initialized exchange_calendar calendar, which is only initialized the first time.
     """
+
     # flag indicating that offset still needs to be checked.
     # A class attribute so we only do this once per class and not per instance
     _FINALIZE_TRADING_CALENDAR = True
@@ -31,12 +35,22 @@ class TradingCalendar(MarketCalendar):
         # offsets of exchange_calendar_mirrors are only available through the instance
         if cls._FINALIZE_TRADING_CALENDAR:
             if self._ec.open_offset:
-                cls.regular_market_times._set("market_open", tuple(
-                    (t[0], t[1], self._ec.open_offset) for t in cls.regular_market_times["market_open"]))
+                cls.regular_market_times._set(
+                    "market_open",
+                    tuple(
+                        (t[0], t[1], self._ec.open_offset)
+                        for t in cls.regular_market_times["market_open"]
+                    ),
+                )
 
             if self._ec.close_offset:
-                cls.regular_market_times._set("market_close", tuple(
-                    (t[0], t[1], self._ec.close_offset) for t in cls.regular_market_times["market_close"]))
+                cls.regular_market_times._set(
+                    "market_close",
+                    tuple(
+                        (t[0], t[1], self._ec.close_offset)
+                        for t in cls.regular_market_times["market_close"]
+                    ),
+                )
             cls._FINALIZE_TRADING_CALENDAR = False
 
         self.__init__(*args, **kwargs)
@@ -85,13 +99,32 @@ class TradingCalendar(MarketCalendar):
     def special_closes_adhoc(self):
         return self._ec.special_closes_adhoc
 
+    @property
+    def weekmask(self):
+        if hasattr(self._ec, "weekmask"):
+            if "1" in self._ec.weekmask or "0" in self._ec.weekmask:
+                # Convert 1s & 0s to Day Abbreviations
+                return " ".join(
+                    [
+                        DAYMASKS[i]
+                        for i, val in enumerate(self._ec.weekmask)
+                        if val == "1"
+                    ]
+                )
+            else:
+                return self._ec.weekmask
+        else:
+            return "Mon Tue Wed Thu Fri"
+
 
 calendars = exchange_calendars.calendar_utils._default_calendar_factories  # noqa
 
-time_props = dict(open_times= "market_open",
-                  close_times= "market_close",
-                  break_start_times= "break_start",
-                  break_end_times= "break_end")
+time_props = dict(
+    open_times="market_open",
+    close_times="market_close",
+    break_start_times="break_start",
+    break_end_times="break_end",
+)
 
 for exchange in calendars:
     cal = calendars[exchange]
@@ -100,15 +133,17 @@ for exchange in calendars:
     regular_market_times = {}
     for prop, new in time_props.items():
         times = getattr(cal, prop)
-        if times is None or isinstance(times, property): continue
+        if times is None or isinstance(times, property):
+            continue
         regular_market_times[new] = times
 
-    cal = type(exchange, (TradingCalendar,), {'_ec_class': calendars[exchange],
-                                              'alias': [exchange],
-                                              'regular_market_times': regular_market_times})
-    locals()[f'{exchange}ExchangeCalendar'] = cal
-
-
-
-
-
+    cal = type(
+        exchange,
+        (TradingCalendar,),
+        {
+            "_ec_class": calendars[exchange],
+            "alias": [exchange],
+            "regular_market_times": regular_market_times,
+        },
+    )
+    locals()[f"{exchange}ExchangeCalendar"] = cal