+++ pandas_market_calendars/holidays/oz.py
@@ -47,7 +47,6 @@ Christmas = Holiday(
     observance=weekend_to_monday,
 )
 
-
 # Boxing day
 BoxingDay = Holiday(
     "Boxing Day",
@@ -56,10 +55,9 @@ BoxingDay = Holiday(
     observance=next_monday_or_tuesday,
 )
 
-
 # One-off holiday additions and removals in Australia
 
 UniqueCloses = []
 
 # National Day of Mourning for Her Majesty the Queen
-UniqueCloses.append(Timestamp("2022-09-22", tz='UTC'))
+UniqueCloses.append(Timestamp("2022-09-22", tz="UTC"))